requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Jinja2==3.1.4
MarkupSafe==3.0.2
Werkzeug==3.1.3
itsdangerous==2.2.0
SQLAlchemy==2.0.36
psycopg==3.2.3
psycopg-binary==3.2.3
WTForms==3.2.1
Mako==1.3.8
alembic==1.14.0
python-dotenv==1.0.1